fixed code structure
