--- SRC/Screens/Signup/Signup_AccountCreated.js
@@ -1,12 +1,0 @@
-import { View, Text } from 'react-native'
-import React from 'react'
-
-const Signup_AccountCreated = () => {
-  return (
-    <View>
-      <Text>Signup_AccountCreated</Text>
-    </View>
-  )
-}
-
-export default Signup_AccountCreated
