+++ SRC/screens/ChatScreen.js
@@ -1,11 +1,12 @@
 import { useEffect } from 'react';
 import { ImageBackground, StyleSheet, FlatList, KeyboardAvoidingView } from 'react-native';
 import { useRoute, useNavigation } from '@react-navigation/native';
-import Message from '../components/Message';
-import InputBox from '../components/InputBox';
+// import Message from '../components/Message';
+import Message from '../../components/Message';
+import InputBox from '../../components/InputBox';
 
-import bg from '../assets/images/BG.png';
-import messages from '../assets/data/messages.json';
+import bg from '../../assets/images/BG.png';
+import messages from '../../assets/data/messages.json';
 
 const ChatScreen = () => {
   const route = useRoute();
