+++ SRC/screens/ChatsScreen.js
@@ -1,6 +1,6 @@
 import { FlatList } from 'react-native';
-import chats from '../assets/data/chats.json';
-import ChatListItem from '../components/ChatListItem';
+import chats from '../../assets/data/chats.json';
+import ChatListItem from '../../components/ChatListItem';
 
 const ChatsScreen = () => {
   return (
