+++ SRC/screens/ContactsScreen.js
@@ -1,6 +1,6 @@
 import { FlatList } from 'react-native';
-import chats from '../assets/data/chats.json';
-import ContactListItem from '../components/ContactListItem';
+import chats from '../../assets/data/chats.json';
+import ContactListItem from '../../components/ContactListItem';
 
 const ContactsScreen = () => {
   return (
