+++ navigation/index.js
@@ -1,9 +1,14 @@
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 
-import ChatScreen from '../screens/ChatScreen';
-import ContactsScreen from '../screens/ContactsScreen';
-import MainTabNavigator from './MainTabNavigator';
+import ChatScreen from '../src/screens/ChatScreen';
+import ContactsScreen from '../src/screens/ContactsScreen';
+// import MainTabNavigator from '../../navigation/MainTabNavigator';
+import MainTabNavigator from '../navigation/MainTabNavigator.js';
+import Login from '../src/screens/Login';
+import Welcome from '../src/screens/Welcome';
+import Signup from '../src/screens/Signup';
+
 
 const Stack = createNativeStackNavigator();
 
@@ -11,6 +16,9 @@ const Navigator = () => {
   return (
     <NavigationContainer>
       <Stack.Navigator screenOptions={{ headerStyle: { backgroundColor: 'whitesmoke' } }}>
+        <Stack.Screen name="Welcome" component={Welcome} options={{ headerShown: false }} />
+        <Stack.Screen name="Login" component={Login} />
+        <Stack.Screen name="Signup" component={Signup} />
         <Stack.Screen name="Home" component={MainTabNavigator} options={{ headerShown: false }} />
         <Stack.Screen name="Chat" component={ChatScreen} />
         <Stack.Screen name="Contacts" component={ContactsScreen} />
